Change %d to %ld and cast to long, which is safer than casting 64-bit values to 32-bit int. Fix some int/magma_int_t compatability. (in testing)
git-svn-id: https://icl.cs.utk.edu/svn/magma/trunk@9360 02bb5f2a-90ca-11de-bb24-8aedcc835bc8

+++ testing/testing_auxiliary.cpp
@@ -48,57 +48,57 @@ void test_num_gpus()
     
     unsetenv("MAGMA_NUM_GPUS");
     ngpu = magma_num_gpus();
-    printf( "%-18s  %7d  %6d\n\n", "not set", ngpu, 1 );
+    printf( "%-18s  %7ld  %6ld\n\n", "not set", long(ngpu), long(1) );
     warn( ngpu == 1 );
     
     setenv("MAGMA_NUM_GPUS", "", 1 );
     ngpu = magma_num_gpus();
-    printf( "%-18s  %7d  %6d\n\n", getenv("MAGMA_NUM_GPUS"), ngpu, 1 );
+    printf( "%-18s  %7ld  %6ld\n\n", getenv("MAGMA_NUM_GPUS"), long(ngpu), long(1) );
     warn( ngpu == 1 );
     
     setenv("MAGMA_NUM_GPUS", "-1", 1 );
     ngpu = magma_num_gpus();
-    printf( "%-18s  %7d  %6d\n\n", getenv("MAGMA_NUM_GPUS"), ngpu, 1 );
+    printf( "%-18s  %7ld  %6ld\n\n", getenv("MAGMA_NUM_GPUS"), long(ngpu), long(1) );
     warn( ngpu == 1 );
     
     setenv("MAGMA_NUM_GPUS", "2junk", 1 );
     ngpu = magma_num_gpus();
-    printf( "%-18s  %7d  %6d\n\n", getenv("MAGMA_NUM_GPUS"), ngpu, 1 );
+    printf( "%-18s  %7ld  %6ld\n\n", getenv("MAGMA_NUM_GPUS"), long(ngpu), long(1) );
     warn( ngpu == 1 );
     
     setenv("MAGMA_NUM_GPUS", "0", 1 );
     ngpu = magma_num_gpus();
-    printf( "%-18s  %7d  %6d\n\n", getenv("MAGMA_NUM_GPUS"), ngpu, 1 );
+    printf( "%-18s  %7ld  %6ld\n\n", getenv("MAGMA_NUM_GPUS"), long(ngpu), long(1) );
     warn( ngpu == 1 );
     
     setenv("MAGMA_NUM_GPUS", "1", 1 );
     ngpu = magma_num_gpus();
-    printf( "%-18s  %7d  %6d\n\n", getenv("MAGMA_NUM_GPUS"), ngpu, 1 );
+    printf( "%-18s  %7ld  %6ld\n\n", getenv("MAGMA_NUM_GPUS"), long(ngpu), long(1) );
     warn( ngpu == 1 );
     
     setenv("MAGMA_NUM_GPUS", "2", 1 );
     ngpu = magma_num_gpus();
-    printf( "%-18s  %7d  %6d\n\n", getenv("MAGMA_NUM_GPUS"), ngpu, 2 );
+    printf( "%-18s  %7ld  %6ld\n\n", getenv("MAGMA_NUM_GPUS"), long(ngpu), long(2) );
     warn( ngpu == min(  2, maxgpu ) );
     
     setenv("MAGMA_NUM_GPUS", "4", 1 );
     ngpu = magma_num_gpus();
-    printf( "%-18s  %7d  %6d\n\n", getenv("MAGMA_NUM_GPUS"), ngpu, 4 );
+    printf( "%-18s  %7ld  %6ld\n\n", getenv("MAGMA_NUM_GPUS"), long(ngpu), long(4) );
     warn( ngpu == min(  4, maxgpu ) );
     
     setenv("MAGMA_NUM_GPUS", "8", 1 );
     ngpu = magma_num_gpus();
-    printf( "%-18s  %7d  %6d\n\n", getenv("MAGMA_NUM_GPUS"), ngpu, 8 );
+    printf( "%-18s  %7ld  %6ld\n\n", getenv("MAGMA_NUM_GPUS"), long(ngpu), long(8) );
     warn( ngpu == min(  8, maxgpu ) );
     
     setenv("MAGMA_NUM_GPUS", "16", 1 );
     ngpu = magma_num_gpus();
-    printf( "%-18s  %7d  %6d\n\n", getenv("MAGMA_NUM_GPUS"), ngpu, 16 );
+    printf( "%-18s  %7ld  %6ld\n\n", getenv("MAGMA_NUM_GPUS"), long(ngpu), long(16) );
     warn( ngpu == min( 16, maxgpu ) );
     
     setenv("MAGMA_NUM_GPUS", "1000", 1 );
     ngpu = magma_num_gpus();
-    printf( "%-18s  %7d  %6d (maxgpu)\n\n", getenv("MAGMA_NUM_GPUS"), ngpu, maxgpu );
+    printf( "%-18s  %7ld  %6ld (maxgpu)\n\n", getenv("MAGMA_NUM_GPUS"), long(ngpu), long(maxgpu) );
     warn( ngpu == min( 1000, maxgpu ) );
     
 #endif // not Windows
@@ -113,38 +113,38 @@ void test_num_threads()
     // test that getting & setting numthreads works
     magma_int_t p_nthread_orig = magma_get_parallel_numthreads();
     magma_int_t l_nthread_orig = magma_get_lapack_numthreads();
-    printf( "get;      parallel_numthread=%2d, lapack_numthread=%2d\n",
-            p_nthread_orig, l_nthread_orig );
+    printf( "get;      parallel_numthread=%2ld, lapack_numthread=%2ld\n",
+            long(p_nthread_orig), long(l_nthread_orig) );
     
     magma_set_lapack_numthreads( 4 );
     magma_int_t p_nthread = magma_get_parallel_numthreads();
     magma_int_t l_nthread = magma_get_lapack_numthreads();
-    printf( "set( 4);  parallel_numthread=%2d, lapack_numthread=%2d (expect  4)\n",
-            p_nthread, l_nthread );
+    printf( "set( 4);  parallel_numthread=%2ld, lapack_numthread=%2ld (expect  4)\n",
+            long(p_nthread), long(l_nthread) );
     warn( p_nthread == p_nthread_orig );
     warn( l_nthread == 4 );
     
     magma_set_lapack_numthreads( 1 );
     p_nthread = magma_get_parallel_numthreads();
     l_nthread = magma_get_lapack_numthreads();
-    printf( "set( 1);  parallel_numthread=%2d, lapack_numthread=%2d (expect  1)\n",
-            p_nthread, l_nthread );
+    printf( "set( 1);  parallel_numthread=%2ld, lapack_numthread=%2ld (expect  1)\n",
+            long(p_nthread), long(l_nthread) );
     warn( p_nthread == p_nthread_orig );
     warn( l_nthread == 1 );
     
     magma_set_lapack_numthreads( 8 );
     p_nthread = magma_get_parallel_numthreads();
     l_nthread = magma_get_lapack_numthreads();
-    printf( "set( 8);  parallel_numthread=%2d, lapack_numthread=%2d (expect  8)\n",
-            p_nthread, l_nthread );
+    printf( "set( 8);  parallel_numthread=%2ld, lapack_numthread=%2ld (expect  8)\n",
+            long(p_nthread), long(l_nthread) );
     warn( p_nthread == p_nthread_orig );
     warn( l_nthread == 8 );
     
     magma_set_lapack_numthreads( l_nthread_orig );
     p_nthread = magma_get_parallel_numthreads();
     l_nthread = magma_get_lapack_numthreads();
-    printf( "set(%2d);  parallel_numthread=%2d, lapack_numthread=%2d (expect %2d)\n",
-            l_nthread_orig, p_nthread, l_nthread, l_nthread_orig );
+    printf( "set(%2ld);  parallel_numthread=%2ld, lapack_numthread=%2ld (expect %2ld)\n",
+            long(l_nthread_orig), long(p_nthread), long(l_nthread), long(l_nthread_orig) );
     warn( p_nthread == p_nthread_orig );
     warn( l_nthread == l_nthread_orig );
     
@@ -162,64 +162,64 @@ void test_num_threads()
         omp_threads = atoi( omp_str );
     }
     
-    printf( "\nusing ncores=%d, omp_num_threads=%d\n\n", ncores, omp_threads );
+    printf( "\nusing ncores=%ld, omp_num_threads=%ld\n\n", long(ncores), long(omp_threads) );
     
     printf( "$MAGMA_NUM_THREADS  nthread  expect\n" );
     printf( "%%==================================\n" );
     
     unsetenv("MAGMA_NUM_THREADS");
     p_nthread = magma_get_parallel_numthreads();
-    printf( "%-18s  %7d  %6d (omp_threads)\n\n", "not set", p_nthread, omp_threads );
+    printf( "%-18s  %7ld  %6ld (omp_threads)\n\n", "not set", long(p_nthread), long(omp_threads) );
     warn( p_nthread == omp_threads );
     
     setenv("MAGMA_NUM_THREADS", "", 1 );
     p_nthread = magma_get_parallel_numthreads();
-    printf( "%-18s  %7d  %6d\n\n", getenv("MAGMA_NUM_THREADS"), p_nthread, 1 );
+    printf( "%-18s  %7ld  %6ld\n\n", getenv("MAGMA_NUM_THREADS"), long(p_nthread), long(1) );
     warn( p_nthread == 1 );
     
     setenv("MAGMA_NUM_THREADS", "-1", 1 );
     p_nthread = magma_get_parallel_numthreads();
-    printf( "%-18s  %7d  %6d\n\n", getenv("MAGMA_NUM_THREADS"), p_nthread, 1 );
+    printf( "%-18s  %7ld  %6ld\n\n", getenv("MAGMA_NUM_THREADS"), long(p_nthread), long(1) );
     warn( p_nthread == 1 );
     
     setenv("MAGMA_NUM_THREADS", "2junk", 1 );
     p_nthread = magma_get_parallel_numthreads();
-    printf( "%-18s  %7d  %6d\n\n", getenv("MAGMA_NUM_THREADS"), p_nthread, 1 );
+    printf( "%-18s  %7ld  %6ld\n\n", getenv("MAGMA_NUM_THREADS"), long(p_nthread), long(1) );
     warn( p_nthread == 1 );
     
     setenv("MAGMA_NUM_THREADS", "0", 1 );
     p_nthread = magma_get_parallel_numthreads();
-    printf( "%-18s  %7d  %6d\n\n", getenv("MAGMA_NUM_THREADS"), p_nthread, 1 );
+    printf( "%-18s  %7ld  %6ld\n\n", getenv("MAGMA_NUM_THREADS"), long(p_nthread), long(1) );
     warn( p_nthread == 1 );
     
     setenv("MAGMA_NUM_THREADS", "1", 1 );
     p_nthread = magma_get_parallel_numthreads();
-    printf( "%-18s  %7d  %6d\n\n", getenv("MAGMA_NUM_THREADS"), p_nthread, 1 );
+    printf( "%-18s  %7ld  %6ld\n\n", getenv("MAGMA_NUM_THREADS"), long(p_nthread), long(1) );
     warn( p_nthread == 1 );
     
     setenv("MAGMA_NUM_THREADS", "2", 1 );
     p_nthread = magma_get_parallel_numthreads();
-    printf( "%-18s  %7d  %6d\n\n", getenv("MAGMA_NUM_THREADS"), p_nthread, 2 );
+    printf( "%-18s  %7ld  %6ld\n\n", getenv("MAGMA_NUM_THREADS"), long(p_nthread), long(2) );
     warn( p_nthread == min(  2, ncores ) );
     
     setenv("MAGMA_NUM_THREADS", "4", 1 );
     p_nthread = magma_get_parallel_numthreads();
-    printf( "%-18s  %7d  %6d\n\n", getenv("MAGMA_NUM_THREADS"), p_nthread, 4 );
+    printf( "%-18s  %7ld  %6ld\n\n", getenv("MAGMA_NUM_THREADS"), long(p_nthread), long(4) );
     warn( p_nthread == min(  4, ncores ) );
     
     setenv("MAGMA_NUM_THREADS", "8", 1 );
     p_nthread = magma_get_parallel_numthreads();
-    printf( "%-18s  %7d  %6d\n\n", getenv("MAGMA_NUM_THREADS"), p_nthread, 8 );
+    printf( "%-18s  %7ld  %6ld\n\n", getenv("MAGMA_NUM_THREADS"), long(p_nthread), long(8) );
     warn( p_nthread == min(  8, ncores ) );
     
     setenv("MAGMA_NUM_THREADS", "16", 1 );
     p_nthread = magma_get_parallel_numthreads();
-    printf( "%-18s  %7d  %6d\n\n", getenv("MAGMA_NUM_THREADS"), p_nthread, 16 );
+    printf( "%-18s  %7ld  %6ld\n\n", getenv("MAGMA_NUM_THREADS"), long(p_nthread), long(16) );
     warn( p_nthread == min( 16, ncores ) );
     
     setenv("MAGMA_NUM_THREADS", "1000", 1 );
     p_nthread = magma_get_parallel_numthreads();
-    printf( "%-18s  %7d  %6d (ncores)\n\n", getenv("MAGMA_NUM_THREADS"), p_nthread, ncores );
+    printf( "%-18s  %7ld  %6ld (ncores)\n\n", getenv("MAGMA_NUM_THREADS"), long(p_nthread), long(ncores) );
     warn( p_nthread == min( 1000, ncores ) );
 #endif // not Windows
 }
@@ -253,7 +253,7 @@ int main( int argc, char** argv )
     test_xerbla();
     
     if ( gFailures > 0 ) {
-        printf( "\n%d tests failed.\n", gFailures );
+        printf( "\n%ld tests failed.\n", long(gFailures) );
     }
     else {
         printf( "\nAll tests passed.\n" );
